requirements.txt
python-dotenv
requests
pillow
openai
google-generativeai
